Migrate Microsoft example DAGs to new design #22452 - mssql
