+++ docs/apache-airflow-providers-microsoft-mssql/operators.rst
@@ -38,7 +38,7 @@ The code snippets below are based on Airflow-2.2
 
 An example usage of the MsSqlOperator is as follows:
 
-.. exampleinclude:: /../../airflow/providers/microsoft/mssql/example_dags/example_mssql.py
+.. exampleinclude:: /../../tests/system/providers/microsoft/mssql/example_mssql.py
     :language: python
     :start-after: [START howto_operator_mssql]
     :end-before: [END howto_operator_mssql]
@@ -46,7 +46,7 @@ An example usage of the MsSqlOperator is as follows:
 You can also use an external file to execute the SQL commands. Script folder must be at the same level as DAG.py file.
 This way you can easily maintain the SQL queries separated from the code.
 
-.. exampleinclude:: /../../airflow/providers/microsoft/mssql/example_dags/example_mssql.py
+.. exampleinclude:: /../../tests/system/providers/microsoft/mssql/example_mssql.py
     :language: python
     :start-after: [START mssql_operator_howto_guide_create_table_mssql_from_external_file]
     :end-before: [END mssql_operator_howto_guide_create_table_mssql_from_external_file]
@@ -68,7 +68,7 @@ Inserting data into a MSSQL database table
 ---------------------------------------------
 We can then create a MsSqlOperator task that populate the ``Users`` table.
 
-.. exampleinclude:: /../../airflow/providers/microsoft/mssql/example_dags/example_mssql.py
+.. exampleinclude:: /../../tests/system/providers/microsoft/mssql/example_mssql.py
     :language: python
     :start-after: [START mssql_operator_howto_guide_populate_user_table]
     :end-before: [END mssql_operator_howto_guide_populate_user_table]
@@ -79,7 +79,7 @@ Fetching records from your MSSQL database table
 
 Fetching records from your MSSQL database table can be as simple as:
 
-.. exampleinclude:: /../../airflow/providers/microsoft/mssql/example_dags/example_mssql.py
+.. exampleinclude:: /../../tests/system/providers/microsoft/mssql/example_mssql.py
     :language: python
     :start-after: [START mssql_operator_howto_guide_get_all_countries]
     :end-before: [END mssql_operator_howto_guide_get_all_countries]
@@ -93,7 +93,7 @@ SQL requests during runtime.
 
 To find the countries in Asian continent:
 
-.. exampleinclude:: /../../airflow/providers/microsoft/mssql/example_dags/example_mssql.py
+.. exampleinclude:: /../../tests/system/providers/microsoft/mssql/example_mssql.py
     :language: python
     :start-after: [START mssql_operator_howto_guide_params_passing_get_query]
     :end-before: [END mssql_operator_howto_guide_params_passing_get_query]
@@ -104,7 +104,7 @@ The complete MSSQL Operator DAG
 
 When we put everything together, our DAG should look like this:
 
-.. exampleinclude:: /../../airflow/providers/microsoft/mssql/example_dags/example_mssql.py
+.. exampleinclude:: /../../tests/system/providers/microsoft/mssql/example_mssql.py
     :language: python
     :start-after: [START mssql_operator_howto_guide]
     :end-before: [END mssql_operator_howto_guide]
