+++ tests/system/providers/microsoft/mssql/create_table.sql
